Update Image model

+++ src/api/Models/Image.cs
@@ -14,7 +14,7 @@ public class Image
     public int? CameraPositionNumber { get; set; }
     public string? CameraPositionName { get; set; }
     public string? WeatherPrediction { get; set; }
-    public float? WeatherPredictionPercent { get; set; }
+    public double? WeatherPredictionPercent { get; set; }
     public string? SnowPrediction { get; set; }
-    public float? SnowPredictionPercent { get; set;}
+    public double? SnowPredictionPercent { get; set; }
 }
